+++ skawa_material_components/lib/card/card.dart
@@ -141,5 +141,5 @@ class SkawaCardHeaderComponent {
     return null;
   }
 
-  static final RegExp _rgbaRegexp = new RegExp(r'rgba?\s*\((?:\d+(?:\.[\d]+)?,?\s*){3,4}\)');
+  static final RegExp _rgbaRegexp = RegExp(r'rgba?\s*\((?:\d+(?:\.[\d]+)?,?\s*){3,4}\)');
 }